Reworked the routing
